[rom_ext,dice] Move dice_chain code from rom_ext to lib/cert
To share the dice code between (imm) rom_ext, those code in rom_ext.c are
refactored and move to the lib/cert/dice_chain* as a common library.

PR #24980 has already refactored the code, and this PR moves them out from
the rom_ext.c to the library folder.

This PR only moves the code with necessary headers / dependencies rewrite.
All functionalities are unchanged.

(cherry picked from commit f682e330a057a99ce5f1759394b5232fcc6c926b)
