+++ sw/device/silicon_creator/rom_ext/BUILD
@@ -177,31 +177,25 @@ cc_test(
             "//sw/device/lib/arch:device",
             "//sw/device/lib/base:csr",
             "//sw/device/lib/base:macros",
+            "//sw/device/lib/base:memory",
             "//sw/device/lib/base:stdasm",
-            "//sw/device/lib/crypto/drivers:entropy",
             "//sw/device/lib/runtime:hart",
-            "//sw/device/silicon_creator/lib:attestation",
             "//sw/device/silicon_creator/lib:boot_data",
             "//sw/device/silicon_creator/lib:boot_log",
             "//sw/device/silicon_creator/lib:dbg_print",
-            "//sw/device/silicon_creator/lib:keymgr_binding",
             "//sw/device/silicon_creator/lib:manifest",
             "//sw/device/silicon_creator/lib:manifest_def",
-            "//sw/device/silicon_creator/lib:otbn_boot_services",
             "//sw/device/silicon_creator/lib:shutdown",
             "//sw/device/silicon_creator/lib/base:chip",
             "//sw/device/silicon_creator/lib/base:sec_mmio",
             "//sw/device/silicon_creator/lib/base:static_critical",
             "//sw/device/silicon_creator/lib/base:util",
             "//sw/device/silicon_creator/lib/boot_svc:boot_svc_msg",
-            "//sw/device/silicon_creator/lib/cert",
-            "//sw/device/silicon_creator/lib/cert:cdi_0_template_library",
-            "//sw/device/silicon_creator/lib/cert:cdi_1_template_library",
+            "//sw/device/silicon_creator/lib/cert:dice_chain",
             "//sw/device/silicon_creator/lib/drivers:ast",
             "//sw/device/silicon_creator/lib/drivers:flash_ctrl",
             "//sw/device/silicon_creator/lib/drivers:hmac",
             "//sw/device/silicon_creator/lib/drivers:ibex",
-            "//sw/device/silicon_creator/lib/drivers:keymgr",
             "//sw/device/silicon_creator/lib/drivers:lifecycle",
             "//sw/device/silicon_creator/lib/drivers:otp",
             "//sw/device/silicon_creator/lib/drivers:pinmux",
@@ -212,7 +206,6 @@ cc_test(
             "//sw/device/silicon_creator/lib/ownership:ownership_activate",
             "//sw/device/silicon_creator/lib/ownership:ownership_unlock",
             "//sw/device/silicon_creator/lib/sigverify",
-            "//sw/device/silicon_creator/manuf/base:perso_tlv_data",
             "//sw/otbn/crypto:boot",
         ] + variation_deps,
     )
